fix issues identified by rtsan and update juce (#66)

+++ source/DSP/PhatEffectsCrossfadeProcessor.hpp
@@ -114,21 +114,21 @@ class EffectsCrossfadeProcessor
                   const juce::dsp::ProcessContextReplacing<T>& context)
     {
         const auto inputBlock { context.getInputBlock() };
-        jassert (previousEffectBuffer.getNumChannels() >= nextEffectBuffer.getNumChannels()
-                 && nextEffectBuffer.getNumChannels() >= inputBlock.getNumChannels());
-        jassert (previousEffectBuffer.getNumSamples () >= nextEffectBuffer.getNumSamples ()
-                 && nextEffectBuffer.getNumSamples () >= inputBlock.getNumSamples ());
+        jassert ((int) previousEffectBuffer.getNumChannels() >= nextEffectBuffer.getNumChannels()
+                 && (int) nextEffectBuffer.getNumChannels() >= inputBlock.getNumChannels());
+        jassert ((int) previousEffectBuffer.getNumSamples() >= nextEffectBuffer.getNumSamples()
+                 && (int) nextEffectBuffer.getNumSamples() >= inputBlock.getNumSamples());
 
         const bool needToInverse = juce::approximatelyEqual (smoothedGainL.getTargetValue (), static_cast<T> (1));
 
         T curGain {};
-        for (int channel = 0; channel < inputBlock.getNumChannels (); ++channel)
+        for (int channel = 0; channel < (int) inputBlock.getNumChannels (); ++channel)
         {
             const auto* prevData = previousEffectBuffer.getReadPointer (channel);
             const auto* nextData = nextEffectBuffer.getReadPointer (channel);
             auto*       outData  = context.getOutputBlock().getChannelPointer (channel);
 
-            for (int sample = 0; sample < inputBlock.getNumSamples(); ++sample)
+            for (int sample = 0; sample < (int) inputBlock.getNumSamples(); ++sample)
             {
                 //TODO VB: this could be an IIFE to get curGain
                 //figure out curGain value based on the current channel and whether we're running the smoothedGains in reverse
